# If applied, this commit will... Fix creating response error
# Explain why this change is being made
When user tried to submit a new response in Endpoint Editor
We were pulling endpoint_id from params. Unfortunately we've had
two instances of <Modals /> component in production. And the second one
did not have props set.

This was a mistake as we shouldn't have <Modals /> added twice. But this bug
was unnoticed because both modals were displaying at top of another.

# Provide links to any relevant tickets, articles or other resources
https://exlabs.atlassian.net/browse/TUT-386